Rename TC versions plugin file & run prettier
