+++ packages/react-scripts/utils/trunkclubVersionsPlugin.js
@@ -1,8 +1,10 @@
-const path = require('path');
+'use strict'
+
+const path = require('path')
 
 function getVersions(packagePath, modulesPath) {
   try {
-    const pkg = require(packagePath);
+    const pkg = require(packagePath)
     const pkgNames = [].concat(
       Object.keys(pkg.dependencies || {}),
       Object.keys(pkg.devDependencies || {})
@@ -10,45 +12,47 @@ function getVersions(packagePath, modulesPath) {
     const versions = pkgNames
       .filter(name => name.startsWith('@trunkclub/'))
       .reduce((acc, name) => {
-        const modulePath = path.join(modulesPath, name, 'package.json');
+        const modulePath = path.join(modulesPath, name, 'package.json')
         try {
           return Object.assign({}, acc, {
-            [name]: require(modulePath).version
-          });
+            [name]: require(modulePath).version,
+          })
         } catch (e) {
-          return acc;
+          return acc
         }
       }, {})
-    return Object.assign({}, { [pkg.name]: pkg.version }, versions);
+    return Object.assign({}, { [pkg.name]: pkg.version }, versions)
   } catch (e) {
-    return { error: `${e.name}: ${e.message}` };
+    return { error: `${e.name}: ${e.message}` }
   }
 }
 
 function TrunkClubVersionsPlugin(options) {
   if (!options || !options.packagePath || !options.modulesPath) {
-    throw new Error('Missing config for TrunkClubVersionsPlugin. Options `packagePath` and `modulesPath` are required.')
+    throw new Error(
+      'Missing config for TrunkClubVersionsPlugin. Options `packagePath` and `modulesPath` are required.'
+    )
   }
 
-  this.versionData = getVersions(options.packagePath, options.modulesPath);
+  this.versionData = getVersions(options.packagePath, options.modulesPath)
 }
 
 TrunkClubVersionsPlugin.prototype.apply = function(compiler) {
-  const file = JSON.stringify(this.versionData);
+  const file = JSON.stringify(this.versionData)
 
   compiler.plugin('emit', function(compilation, callback) {
     // Add a file to the webpack build
     compilation.assets['tcversions.json'] = {
       source() {
-        return file;
+        return file
       },
       size() {
-        return file.length;
-      }
-    };
+        return file.length
+      },
+    }
 
-    callback();
-  });
-};
+    callback()
+  })
+}
 
-module.exports = TrunkClubVersionsPlugin;
+module.exports = TrunkClubVersionsPlugin
